feat: 각 View들의 updateView함수에 인자 추가
