+++ burstcamp/burstcamp/Scene/Common/View/DefaultBadgeView.swift
@@ -53,10 +53,17 @@ final class DefaultBadgeView: UIView {
 }
 
 extension DefaultBadgeView {
-    func updateView(user: MyUser) {
+    func updateView(user: User) {
         domainLabel.updateView(text: user.domain.rawValue)
         let number = "\(user.ordinalNumber)기"
         numberLabel.updateView(text: number)
         camperIDLabel.updateView(text: user.camperID)
+    }
+
+    func updateView(feedWriter: FeedWriter) {
+        domainLabel.updateView(text: feedWriter.domain.rawValue)
+        let number = "\(feedWriter.ordinalNumber)기"
+        numberLabel.updateView(text: number)
+        camperIDLabel.updateView(text: feedWriter.camperID)
     }
 }
